backend/requirements.txt
Django
psycopg2
django-debug-toolbar
djangorestframework
django-cors-headers
django-filter
djangorestframework
gunicorn